Recent Order functionality

+++ server/db/Models.js
@@ -14,7 +14,7 @@ const Investor = db.define(
     InvestorEmail: DataTypes.STRING,
     InvestorPassword: DataTypes.TEXT,
     Username: DataTypes.TEXT,
-    NetWorth: DataTypes.BIGINT,
+    NetWorth: DataTypes.DECIMAL(20, 2),
     InvestorRanking: DataTypes.INTEGER,
     InvestorDifficulty: DataTypes.TEXT,
     DateJoined: DataTypes.DATEONLY,
